运动量小数点修改

+++ app/src/main/java/com/xiaomaoqiu/now/bussiness/bean/PetInfoBean.java
@@ -45,7 +45,7 @@ public class PetInfoBean extends BaseBean {
 
 
 
-    public int reality_amount;
+    public double reality_amount;
 
 
 }
